+++ src/Chinchilla.Integration/Features/Messages/HelloWorldMessage.cs
@@ -1,12 +1,21 @@
+using System.Collections.Generic;
+
 namespace Chinchilla.Integration.Features.Messages
 {
-    public class HelloWorldMessage : IHasRoutingKey, IHelloWorldMessage
+    public class HelloWorldMessage : IHasRoutingKey, IHelloWorldMessage, IHasHeaders
     {
         public string Message { get; set; }
 
         string IHasRoutingKey.RoutingKey
         {
             get { return "messages." + Message; }
+        }
+
+        public void PopulateHeaders(IDictionary<object, object> headers)
+        {
+            headers.Add("key1", "foo");
+            headers.Add("key2", "foo");
+            headers.Add("key3", "foo");
         }
     }
 }