+++ src/Chinchilla/IHasHeaders.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Chinchilla
+{
+    public interface IHasHeaders
+    {
+        void PopulateHeaders(IDictionary<object, object> headers);
+    }
+}