+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Pending
 
+ * Custom headers on publish
  * [BUG-#31] [tanir] Fault strategies cause exception when iterating delivery listeners multiple times
 
 # 0.2.1.0